Rename page components
